Continued testing if DRF outputs anything other than None.
DRF Source block in GRC seems to create empty files only with metadata in them. Are samples not being passed through? I'm not sure, but both start and end indices are specified, repeat is True (while throttle and gapless are not), and specification of data filepaths and channels is correct. At least, all this is true to the best of my ability.